refactor(components): shared/ vira FLAT — CHECK 3 no guard + MOVE shared/ponto → Pages/Ponto/_components
Furo achado horas depois do components-tree-guard nascer: subpasta de
dominio DENTRO de shared/ (ponto/) escapava do CHECK 1 (so top-level).
Regra nova enforcavel: shared/ e flat — composto cross-modulo e arquivo
direto; dominio de 1 modulo vai pra Pages/<Mod>/_components/.

- MOVE shared/ponto/{ActivityFeed,AlertInbox,MonthHeatmap,PresenceStrip}
  → Pages/Ponto/_components/ (consumidores: Ponto/Dashboard · Espelho ·
  _Showcase — imports atualizados)
- CHECK 3 no components-tree-guard (sem grandfather)
- Baselines re-keyed: eslint (+6 no-restricted-syntax absorvidos — hand-roll
  pre-existente entrando no escopo DS Pages/**) · layout · ui-lint
- CHANGELOG DS 0.6.12

Gates: components:check ✅ shared/ flat · layout ✅ · lint:baseline ✅ ·
reuse:gate ✅. Zero mudanca visual.

Refs: CYCLE-08 · ADR UI-0013 · ADR 0240 · auditoria duplicatas 2026-06-11
