+++ resources/js/Pages/Ponto/_components/PresenceStrip.tsx
