+++ scripts/layout-primitives-baseline.json
@@ -34,10 +34,10 @@
     "resources/js/Components/shared/PageHeader.tsx": 3,
     "resources/js/Components/shared/PageHeaderActions.tsx": 1,
     "resources/js/Components/shared/PageHeaderTabs.tsx": 1,
-    "resources/js/Components/shared/ponto/ActivityFeed.tsx": 5,
-    "resources/js/Components/shared/ponto/AlertInbox.tsx": 5,
-    "resources/js/Components/shared/ponto/MonthHeatmap.tsx": 6,
-    "resources/js/Components/shared/ponto/PresenceStrip.tsx": 3,
+    "resources/js/Pages/Ponto/_components/ActivityFeed.tsx": 5,
+    "resources/js/Pages/Ponto/_components/AlertInbox.tsx": 5,
+    "resources/js/Pages/Ponto/_components/MonthHeatmap.tsx": 6,
+    "resources/js/Pages/Ponto/_components/PresenceStrip.tsx": 3,
     "resources/js/Components/shared/SubNav.tsx": 1,
     "resources/js/Components/shared/VendaDerivadaCard.tsx": 8,
     "resources/js/Components/Site/DashboardMockup.tsx": 9,
